% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSSA.R
\name{gSSA}
\alias{gSSA}
\title{stochastic simulation function}
\usage{
gSSA(N, R0=NULL, beta=NULL,
                 S0, I0, E0,
                 model=NULL,
                 gamma=0, sigma=0, rho=0, epsilon=0, omega=0,
                 b, m=NULL,
                 whichBeta=NULL, whichSigma=NULL, whichGamma=NULL,
                 submodel=NULL,
                 tmax=10, inc=52)
}
\description{
stochastic simulation function: requires either R0 or beta to be specified, initial compartment populations for S0, I0, and E0, and either a model specification and parameters (gamma, sigma, rho, epsilon, omega) or a submodel object
}
