\name{submodel}
\alias{submodel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
submodel object, indicating model specification and parameters
}
\description{
submodel object, indicating model specification and parameters
}
\usage{
submodel(model, omega, gamma, rho, sigma, epsilon)}

